[PROF-6535] no unnecessary type casting

+++ profiling/src/lib.rs
@@ -313,8 +313,9 @@ pub struct RequestLocals {
 }
 
 /// take a sample every X bytes
+/// this value is temporary but the overhead looks promising, Go profiler samples every 512 KiB
 #[cfg(feature = "allocation_profiling")]
-const ALLOCATION_PROFILING_INTERVAL: f32 = 1024.0 * 50.0;
+const ALLOCATION_PROFILING_INTERVAL: f64 = 1024.0 * 50.0;
 
 #[cfg(feature = "allocation_profiling")]
 pub struct AllocationProfilingStats {
@@ -343,7 +344,7 @@ impl AllocationProfilingStats {
             return;
         }
 
-        let scale = 1.0 / (1.0 - (len as f64 * -1.0 / ALLOCATION_PROFILING_INTERVAL as f64).exp());
+        let scale = 1.0 / (1.0 - (len as f64 * -1.0 / ALLOCATION_PROFILING_INTERVAL).exp());
 
         let count = 1.0 * scale;
         let bytes = len as f64 * scale;
